Revert "Feature/opti utility"
