+++ PuppyEngine/.env
@@ -51,5 +51,3 @@ STORAGE_SERVER_URL=https://dev.storage.puppyagent.com
 verfication_code=203177
 
 DEPLOYMENT_TYPE=Local
-
-PUPPY_ENV=development
