+++ PuppyStorage/.env
@@ -29,5 +29,3 @@ CLOUDFLARE_R2_SECRET_ACCESS_KEY=0b97bf5facc3d4535678e83d36d48a965df523edccb376ca
 SUPABASE_URL=postgresql://postgres.vpgbquumhkqkjkhrbkke:PuppyAgentGTM@aws-0-us-west-1.pooler.supabase.com:6543/postgres?sslmode=require
 
 STORAGE_TYPE=Local
-
-PUPPY_ENV=development
