+++ PuppyStorage/server/storage_server.py
@@ -6,28 +6,14 @@ from fastapi import FastAPI, Request
 from fastapi.responses import JSONResponse
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.exceptions import RequestValidationError
-from tools.puppy_utils import PuppyException
-from tools.puppy_utils import log_info, log_error, log_warning
-
-# 获取环境信息并创建相应的日志器
-from tools.puppy_utils.logger import get_logger
-from tools.puppy_utils.config import ENV
-
-# 获取特定服务的日志器
-storage_logger = get_logger("puppystorage")
-log_info = storage_logger.info
-log_warning = storage_logger.warning
-log_error = storage_logger.error
-
+from utils.puppy_exception import PuppyException
+from utils.logger import log_info, log_error
 from server.routes.file_routes import file_router, storage_router
 from server.routes.vector_routes import vector_router
 
 
 try:
     app = FastAPI()
-
-    # 记录当前环境信息
-    log_info(f"Storage server starting in {ENV} environment")
 
     # Add CORS middleware
     app.add_middleware(
@@ -51,8 +37,7 @@ try:
         raise PuppyException(
             error_code=3000,  # Validation error code
             error_message="Validation Error",
-            cause=error_message,
-            service_name="puppystorage"
+            cause=error_message
         )
 
     app.include_router(vector_router)
