+++ PuppyStorage/utils/puppy_exception.py
@@ -1,20 +1,50 @@
-"""
-这是一个重定向文件，用于向后兼容。
-推荐直接使用 tools.puppy_utils 包中的异常类和处理函数。
-"""
+import traceback
+from functools import wraps
+from utils.logger import log_error, log_info
 
-import warnings
+class PuppyException(Exception):
+    service_name = "puppystorage"  
+    
+    def __init__(
+        self,
+        error_code: int,
+        error_message: str,
+        cause: str = None
+    ):
+        self.error_code = error_code
+        self.error_message = error_message
+        self.cause = cause
+        self.raise_message = f"[{self.service_name.upper()}_ERROR_{self.error_code}]: {self.error_message}!"
+        if self.cause:
+            self.raise_message += f"\nCause: {self.cause}"
+        super().__init__(self.raise_message)
 
-# 导入新工具包中的异常类和处理器
-from tools.puppy_utils.puppy_exception import PuppyStorageException as NewPuppyException
-from tools.puppy_utils.puppy_exception import global_exception_handler
 
-# 显示弃用警告
-warnings.warn(
-    "PuppyStorage.utils.puppy_exception 已弃用，请使用 tools.puppy_utils 中的异常处理模块",
-    DeprecationWarning,
-    stacklevel=2
-)
-
-# 为了向后兼容，使用别名
-PuppyException = NewPuppyException
+def global_exception_handler(
+    error_code: int,
+    error_message: str,
+    log_at_root: bool = False
+):
+    def decorator(func):
+        @wraps(func)
+        def wrapper(*args, **kwargs):
+            try:
+                return func(*args, **kwargs)
+            except PuppyException as e:
+                # Propagate the error without re-logging
+                if not log_at_root:
+                    raise
+                tb_str = traceback.format_exc()
+                full_error_message = f"{str(e)}\nTraceback:\n{tb_str}"
+                log_error(full_error_message)
+                raise
+            except Exception as e:
+                # Wrap and propagate the exception, logging only at root level
+                if not log_at_root:
+                    raise PuppyException(error_code, error_message, str(e))
+                tb_str = traceback.format_exc()
+                full_error_message = f"[{PuppyException.service_name.upper()}_ERROR_{error_code}]: {error_message}\nCause: {str(e)}\nTraceback:\n{tb_str}"
+                log_error(full_error_message)
+                raise PuppyException(error_code, error_message, str(e))
+        return wrapper
+    return decorator
