--- tools/puppy_utils/__init__.py
@@ -1,16 +1,0 @@
-"""
-Puppy Utilities Package
-
-这个包提供了 PuppyAgent 项目中的公共工具函数和类，
-包括日志记录、配置管理和异常处理等功能。
-"""
-
-from tools.puppy_utils.logger import Logger, log_info, log_error, log_warning, log_debug
-from tools.puppy_utils.config import AppConfig, config, PathManager, paths
-from tools.puppy_utils.puppy_exception import PuppyException, global_exception_handler
-
-__all__ = [
-    'Logger', 'log_info', 'log_error', 'log_warning', 'log_debug',
-    'AppConfig', 'config', 'PathManager', 'paths',
-    'PuppyException', 'global_exception_handler'
-]
